python/requirements.txt
pyyaml>=6.0
certifi>=2024.0.0
cffi>=1.16.0
chardet>=5.2.0
charset-normalizer>=3.3.0
cryptography>=43.0.0
dnspython>=2.6.0
hvac>=2.3.0
idna>=3.7
packaging>=24.0
pycparser>=2.22
pyOpenSSL>=24.0.0
PySocks>=1.7.1
requests>=2.32.0
SOLIDserverRest>=2.3.9
urllib3>=2.2.0
